Split the evidence archive out of the IPC monolith
  electron/ipc.ts     2029-2247 -> electron/ipc/archive.ts     (31 handlers)
  electron/preload.ts   468-499 -> electron/preload/archive.ts (31 bindings)
  shared/types.ts     7812-7856 -> shared/api/archive.ts       (31 signatures)

Unlike the six before it, this block was not pure: three things belonging to other
domains sat inside its range and had to be moved out first, each as its own step —
the two GEDCOM handlers (genealogy) in ipc.ts, and scanWorkRecords (works) in both
preload.ts and the NodusApi declaration. All three were relocated next to their own
domain's code, no behaviour touched.

test-genealogy-archive-entry.mjs read the hot files through
`new URL('../electron/ipc.ts', import.meta.url)`, a form the earlier sweep did not
recognise. It now uses the census sentinels; no test reads those three paths in any
form any more.

Where the split stands: 401 of 1,206 channels live in a module named after their
domain, 805 remain in electron/ipc.ts. That file is 3,654 lines (from 6,013 — down
39%), electron/preload.ts 1,296 (from 1,862 — down 30%), shared/types.ts 8,744 (from
9,553).

Verified: 1,425 tests, lint and typecheck clean, contract test green.
